refactor: remove app-list logic from DiscussionsSettings
This is a ‘contract’ commit which removes functionality from DiscussionsSettings and adds it to AppList.

A word of warning that I tried my best to separate this refactor from that of AppConfigForm, but this commit may not be perfect.  The original problem was that the two sub-modules were unnecessarily intertwined, making use of some shared variables when they didn’t need to (such as selectedAppId), so I can’t fully get app-list working without touching app-config-form as well… so this particular commit probably has logical errors and won’t build properly.  Taken together with the next commit for app-config-form, though, they move a significant amount of logic out of DiscussionsSettings and into the sub-modules and isolate them from each other.

You’ll see here that handleSelectApp and handleStartConfig move into the app-list sub-module.  The former goes into AppList, the latter was committed to AppListNextButton in a prior commit.

Once those two are gone, a bunch of other code in DiscussionsSettings starts falling away.  selectedAppId is still necessary since the config form handlers are using it, but we’ll see that disappear in the next commit.
